Create the the variables and functios necessary to play the game without the for loop
create variable to house the prompt value. create variable to house a tolowercase() method linked to the value of the prompt. create a variable to house a function which picks a random value between 1-3, with each value being linked a certain output using a switch statement. create a function which houses a second function. the second function will take the value outputed by the random value function and the value inputed by the user through the prompt. An if statement compares the arguments to ouput the necessary string.