complete the project with a for loop and final score
add a for loops to play the game for 5 rounds. create two variables to hold the score for "computer" and "player". at the end of each round console.log the result. Make the result award points to either 'player' or computer' and console.log the current scoreboard. At the end of the 5 rounds announce the winner.